requirements.txt
flask
python-dotenv
groq
chromadb
sentence-transformers
